+++ NEWS.md
@@ -4,6 +4,7 @@
 
 ### Changes
 
+* Added option to set a second run-script [[#32](https://github.com/sjPlot/Relaunch64/issues/31)].
 * Added start-option _R64BG_ to force emulator not to block Relaunch64.
 * Make warnings clickable, even if otherwise ignored.
 * For 64tass compiler, warnings now give more detailed information from included files.